+++ packages/shared/src/index.ts
@@ -1,10 +1,12 @@
 export { FlowgladActionKey, HTTPMethod } from './types'
 export {
   createCheckoutSessionSchema,
+  cancelSubscriptionSchema,
   flowgladActionValidators,
 } from './actions'
 export type {
   FlowgladActionValidatorMap,
   CreateCheckoutSessionParams,
+  CancelSubscriptionParams,
 } from './actions'
 export { getBaseURL } from './utils'