+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flowglad/server
 
+## 0.8.10
+
+### Patch Changes
+
+- Flow through output metadata
+- Updated dependencies
+  - @flowglad/shared@0.8.10
+
 ## 0.8.9
 
 ### Patch Changes