+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flowglad/types
 
+## 0.8.10
+
+### Patch Changes
+
+- Flow through output metadata
+
 ## 0.8.9
 
 ### Patch Changes